#2 - Reajustando a Tela de Login e fazendo mudanças minimas na tela de Cadastro.

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('name');
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
-            $table->integer('cpf')->unique();
+            $table->bigInteger('cpf')->unique();
             $table->integer('phone');
             $table->string('password');
             $table->rememberToken();
